CHECKME: gcc 4 fix (lvalue cast). Was "(ULONG)(*string2++) =". What was that cast doing?
git-svn-id: https://svn.aros.org/svn/aros/trunk@23818 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/locale/english.c
@@ -327,7 +327,7 @@ AROS_LH4(ULONG, strconvert,
     {
         while(--length && *string1)
         {
-            (ULONG)(*string2++) = collTab[ (UBYTE)*string1 ];
+            *string2++ = collTab[ (UBYTE)*string1 ];
             count++;
         }
         *string2 = '\0';
